Combine boa_error_handlers and unwrap_or_trap
